Fixed some warnings.

+++ mask.c
@@ -152,7 +152,7 @@ unsigned char *Mask_makeMaskedFrame(int width, unsigned char *frame, int mask)
 {
 	unsigned char *masked;
 
-	masked = (unsigned char *)malloc(width * width);
+	masked = (unsigned char *)malloc((size_t)(width * width));
 	if(masked == NULL) return NULL;
 
 	maskMakers[mask](width, frame, masked);
@@ -170,7 +170,7 @@ unsigned char *Mask_makeMask(int width, unsigned char *frame, int mask, QRecLeve
 		return NULL;
 	}
 
-	masked = (unsigned char *)malloc(width * width);
+	masked = (unsigned char *)malloc((size_t)(width * width));
 	if(masked == NULL) return NULL;
 
 	maskMakers[mask](width, frame, masked);
@@ -329,9 +329,9 @@ unsigned char *Mask_mask(int width, unsigned char *frame, QRecLevel level)
 	int demerit;
 	int w2 = width * width;
 
-	mask = (unsigned char *)malloc(w2);
+	mask = (unsigned char *)malloc((size_t)w2);
 	if(mask == NULL) return NULL;
-	bestMask = (unsigned char *)malloc(w2);
+	bestMask = (unsigned char *)malloc((size_t)w2);
 	if(bestMask == NULL) {
 		free(mask);
 		return NULL;
@@ -349,7 +349,7 @@ unsigned char *Mask_mask(int width, unsigned char *frame, QRecLevel level)
 //		printf("(%d,%d,%d,%d)=%d\n", n1, n2, n3 ,n4, demerit);
 		if(demerit < minDemerit) {
 			minDemerit = demerit;
-			memcpy(bestMask, mask, w2);
+			memcpy(bestMask, mask, (size_t)w2);
 		}
 	}
 	free(mask);
